Refactor the rest of components into relevant subfolders
