+++ src/components/sketch/asset-info/assetinfo.scss
