+++ src/components/sketch/asset-info/index.js
