+++ src/components/sketch/view-asset/index.js
