+++ src/components/sketch/view-asset/viewasset.scss
