+++ src/components/sketch/view-asset/viewassetcontroller.js
