+++ src/pages/homepage/homepage.js
@@ -1,5 +1,5 @@
 import HomeHero from "../../components/layout/home-hero"
-import ProjectsList from "../../components/projects-list"
+import ProjectsList from "../../components/sketch/projects-list"
 import "./homepage.scss"
 
 function Homepage() {
